+++ app/src/main/java/com/zywczas/rickmorty/model/db/CharacterFromDb.kt
@@ -6,10 +6,10 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "characters")
 data class CharacterFromDb (
-//todo jako long
+
     @PrimaryKey
     @ColumnInfo(name = "id")
-    val id: Int,
+    val id: Long,
 
     @ColumnInfo(name = "name")
     val name: String,
